dshUnit: Refactored the following dshUnitUI.sh functions to accept the name of the assertion or test as an optional first argument: showAssertionPassedMsg() showAssertionFailedMsg() showTestPassedMsg() showTestFailedMsg()
This will show the name of the assertion or test that passed or failed instead of just the
word "Assertion" or "Test" in the pass/fail messages output.

This makes debugging easier when many assertions or tests are run at once.

If the first argument is not specified, then the word "Assertion" or
"Test" will still be used.

Also refactored the following dshUnit functions to accept the name of
the assertion or test as an optional first argument.

increasePassingAssertions()
increaseFailedAssertions()
increasePassingTests()
increaseFailingTests()

All dshUnit tests are passing.

This commit is related to issue #26.