DSH: Implemented `helpTests.sh::testDshHelpCreatesSystemManPageForDsh()`
All dsh tests are passing. This commit is related to issue #27.

Also: Ran `man dsh` after runnig dsh tests, the man page displayed as
expected. `dsh --help` is now ready to be refactored to show a proper man
page, and also create the system man page for dsh so man dsh can be used
as well as `dsh --help`.