Working on addressing issue #371.
https://github.com/sevidmusic/roady/issues/371

Modified:

-
[`core/abstractions/component/UserInterface/WebUI.php`](https://github.com/sevidmusic/roady/blob/roady/Tests/Unit/interfaces/component/UserInterface/TestTraits/WebUITestTrait.php)

Finished re-defining expected html
structure for output of WebUI.

All tests are passing again, however,
the WebUI is not actually testing
that the appropriate script tags
for js files and link tags for css
files are being incorporated into
the output.

The WebUI is incorporating the link
tags for the appropriate css files,
but again there are not tests for
this.

script tags for js files are not
incorporated or tested.

[Issue](https://github.com/sevidmusic/roady/issues/373)
the fact that the
[`WebUITestTrait`](https://github.com/sevidmusic/roady/blob/roady/Tests/Unit/interfaces/component/UserInterface/TestTraits/WebUITestTrait.php)
does not yet set up the expectation that `link`
tags and `script` tags for appropriate `css`
and `js` files are incorporated into a `WebUI`'s
output.

Modified:

- `core/abstractions/component/UserInterface/WebUI.php`